refactor: Error 관련 DTO 패키지 이동
